fix: fixed an issue with max_attempts and decay_seconds being received as string from config

+++ config/notification.php
@@ -3,7 +3,7 @@
 return [
 
     'rate_limit' => [
-        'max_attempts' => env("NOTIFICATION_RATE_LIMIT_MAX_ATTEMPTS", 10),
-        'decay_seconds' => env("NOTIFICATION_RATE_LIMIT_DECAY_SECONDS", 3600),
+        'max_attempts' => (int) env("NOTIFICATION_RATE_LIMIT_MAX_ATTEMPTS", 10),
+        'decay_seconds' => (int) env("NOTIFICATION_RATE_LIMIT_DECAY_SECONDS", 3600),
     ],
 ];
